requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pdfplumber>=0.11.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-genai>=0.8.0
langchain-text-splitters>=0.3.0
python-multipart>=0.0.9
qdrant-client>=1.8.0
sentence-transformers>=2.5.0